fixed cart ordertable added payment status modified orderservice

+++ BACKEND/lifelinepathlab/src/main/java/com/lifelinepathlab/service/OrderService.java
@@ -2,6 +2,7 @@ package com.lifelinepathlab.service;
 
 import java.util.Date;
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,43 +16,70 @@ import com.lifelinepathlab.repository.OrderRepository;
 
 @Service
 public class OrderService {
-	
+
 	@Autowired
 	private OrderRepository orderRepository;
+	@Autowired
+	private TestService service;
+
+	public void addOrder(Orders orders) {
+		User user = orders.getUser();
+		int userId = user.getUserId();
+		double totalAmount=0;
+		Orders existingOrder = orderRepository.findByUserId(userId, "AC");
+		List<Test> tests = orders.getTests();
+		
 	
-	public void addorder(Orders orders) {
-		orders.setDate(new Date());
-		orders.setStatus("AC");
-		orderRepository.save(orders);
+		
+		if (existingOrder == null) {
+			
+			for(Test test : tests) {
+				Test t2 = service.getTestById(test.getTestId()).get();
+				totalAmount = t2.getFinalPrice();
+			}
+			orders.setDate(new Date());
+			orders.setStatus("AC");
+			orders.setTotalAmount(totalAmount);
+			orderRepository.save(orders);
+				
+		} else {
+			for(Test test: tests) {
+				Test t = service.getTestById(test.getTestId()).get();
+				System.out.println(existingOrder.getTotalAmount());
+
+				totalAmount =existingOrder.getTotalAmount() +  t.getFinalPrice();
+				System.out.println(totalAmount);
+			}
+			existingOrder.getTests().addAll(tests);
+			existingOrder.setTotalAmount(totalAmount);
+			orderRepository.save(existingOrder);
+			
+		}
+		
+		
+			
 	}
-	
-	public List<Orders> getAllOrders(){
+
+	public List<Orders> getAllOrders() {
 		return orderRepository.findAll();
 	}
-	
-//	  public List<Test> getAllTestsByUserId(int userId) {
-//		  
-//	        List<Orders> userOrders = orderRepository.findByUserId(userId , "AC");
-//
-//	        // Extracting all tests from user orders
-//	        return userOrders.stream()
-//	                .flatMap(order -> order.getTests().stream())
-//	                .collect(Collectors.toList());
-//	    }
-	
-	  public List<Orders> getAllTestsByUserId(int userId) {
-	  
-      List<Orders> userOrders = orderRepository.findByUserId(userId , "AC");
-      return userOrders;
-  }
-	
-	
+
+	public List<Orders> getOrdersByStatus(String status) {
+		return orderRepository.findByStatus(status);
+	}
+
+	public Orders getOrderByUserId(int userId) {
+
+		Orders userOrders = orderRepository.findByUserId(userId, "AC");
+		return userOrders;
+	}
+
 	public Orders getOrdersByid(int id) {
 		return orderRepository.findById(id)
 				.orElseThrow(() -> new ResourceNotFoundException("Order does not exits with Order Id: ", id));
 
 	}
-	
+
 	public void updateOrder(Orders orders, int id) {
 		Orders orders2 = getOrdersByid(id);
 		orders2.setStatus(orders.getStatus());
@@ -59,8 +87,34 @@ public class OrderService {
 		orderRepository.save(orders2);
 	}
 
-	public void deleteOrder(int id) {
-		Orders orders = getOrdersByid(id);
-		orderRepository.delete(orders);
+	public void deleteTestFromOrder(int orderId, int testId) {
+		Orders orders = getOrdersByid(orderId);
+		List<Test> tests = orders.getTests();
+
+		//System.out.println(tests);
+				
+		double finalprice =0;
+		if (orders != null) {
+			Test testToRemove = null;
+			for (Test test : tests) {
+				if (test.getTestId() == testId) {
+					testToRemove = test;
+					finalprice = test.getFinalPrice();
+					break;
+				}
+			}
+			
+
+			if (testToRemove != null) {
+				tests.remove(testToRemove);
+				orders.setTotalAmount(orders.getTotalAmount() - finalprice);
+				orderRepository.save(orders);
+			}
+			
+			if (tests.isEmpty()) {
+				orderRepository.delete(orders);
+			}
+		}
 	}
+
 }
